pyproject.toml
[project]
name = "king-vs-pawn-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask-socketio>=5.5.1",
    "flask>=3.1.3",
    "numpy>=2.4.3",
    "python-chess>=1.999",
]
